reincarnation commit
`magit-svn.el' is no longer part of the Magit repository but
its karma is good and so it is reborn as a repository of its own.

To help it remember its past life do this:

  git remote add magit git://github.com/magit/magit.git
  git config remote.magit.tagopt --no-tags
  git fetch magit
  git replace --graft <this> 1176825ca145e35823000564d3b0bedccd334d05

See git-replace(1) for more information.

While it was still part of the Magit repository `magit-svn.el'
was edited by the following people:

    34  Jonas Bernoulli
     5  Phil Jackson
     5  Yann Hodique
     3  Pieter Praet
     2  Leonardo Etcheverry
     2  Rüdiger Sonderfeld
     1  Dan LaManna
     1  Graham Clark
     1  Laurent Laffont
     1  Marc Herbert
     1  Rémi Vanicat
     1  Thomas Jost