clarify copyright
Previously the library header said that the copyright belongs to "The
Magit Project Developers", but that could be read as "the copyright
belongs to the maintainers", when actually nobody assigned his/her
rights to anyone.  So use "contributors" instead, to indicate that
everyone who made a copyrightable contribution, still holds that
right.